refactor: move getDestinationFlowAction, postValidationFormUserData, validateFormUserData to parent level
