+++ app/routes/shared/formular.server.ts
@@ -6,12 +6,12 @@ import { flows } from "~/domains/flows.server";
 import { retrieveContentData } from "~/services/flow/formular/contentData/retrieveContentData";
 import { isFileUploadOrDeleteAction } from "~/services/flow/formular/fileUpload/isFileUploadOrDeleteAction";
 import { processUserFile } from "~/services/flow/formular/fileUpload/processUserFile.server";
-import { getUserDataAndFlow } from "~/services/flow/formular/userDataAndFlow/getUserDataAndFlow";
-import { getDestinationFlowAction } from "~/services/flow/formular/userFlowAction/getDestinationFlowAction";
-import { postValidationFormUserData } from "~/services/flow/formular/userFlowAction/postValidationFormUserData";
-import { validateFormUserData } from "~/services/flow/formular/userFlowAction/validateFormUserData";
 import { addPageDataToUserData } from "~/services/flow/pageData";
 import { buildFlowController } from "~/services/flow/server/buildFlowController";
+import { getUserDataAndFlow } from "~/services/flow/userDataAndFlow/getUserDataAndFlow";
+import { getDestinationFlowAction } from "~/services/flow/userFlowAction/getDestinationFlowAction";
+import { postValidationFormUserData } from "~/services/flow/userFlowAction/postValidationFormUserData";
+import { validateFormUserData } from "~/services/flow/userFlowAction/validateFormUserData";
 import { logWarning } from "~/services/logging";
 import { validatedSession } from "~/services/security/csrf/validatedSession.server";
 import { getSessionManager, updateSession } from "~/services/session.server";
